+++ src/lib.rs
@@ -1,16 +1,23 @@
 mod project {
     pub(crate) mod bin {}
     mod lib {
+        mod enums {
+            mod file_status_kind;
+
+            pub use file_status_kind::*;
+        }
         mod structs {
             mod author_info;
             mod branches;
             mod commit_info;
+            mod file_status_entry;
             mod message;
             mod repository_info;
 
             pub use author_info::*;
             pub use branches::*;
             pub use commit_info::*;
+            pub use file_status_entry::*;
             pub use message::*;
             pub use repository_info::*;
         }
@@ -35,6 +42,7 @@ mod project {
             }
         }
 
+        pub use enums::*;
         pub use structs::*;
     }