+++ src/project/lib/enums/file_status_kind.rs
@@ -1,0 +1,9 @@
+pub enum FileStatusKind {
+    NewStaged,
+    ModifiedStaged,
+    DeletedStaged,
+    NewWt,
+    ModifiedWt,
+    DeletedWt,
+    Ignored,
+}