+++ src/project/lib/structs/file_status_entry.rs
@@ -1,0 +1,14 @@
+mod new;
+
+use std::path::PathBuf;
+use crate::FileStatusKind;
+
+mod get {
+    mod path;
+    mod status;
+}
+
+pub struct FileStatusEntry {
+    path: PathBuf,
+    status: FileStatusKind,
+}