+++ src/project/lib/structs/file_status_entry/get/path.rs
@@ -1,0 +1,8 @@
+use std::path::PathBuf;
+use crate::FileStatusEntry;
+
+impl FileStatusEntry {
+    pub fn path(&self) -> &PathBuf {
+        &self.path
+    }
+}