+++ src/project/lib/structs/file_status_entry/get/status.rs
@@ -1,0 +1,7 @@
+use crate::{FileStatusEntry, FileStatusKind};
+
+impl FileStatusEntry {
+    pub fn status(&self) -> &FileStatusKind {
+        &self.status
+    }
+}